refactor(table): extract shared cellAttributes helper for TableCell and TableHeader
